Added working syllabus upload feature
