+++ .vscode/settings.json
@@ -1,0 +1,5 @@
+{
+    "python.pythonPath": "/Users/nat/.local/share/virtualenvs/hyperschedule-scraper-nZNHL86x/bin/python",
+    "python.linting.pylintEnabled": true,
+    "python.linting.enabled": true
+}
